requirements.txt
aiosmtplib==2.0.2
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
click==8.1.7
dnspython==2.6.1
docx2txt==0.8
email-validator==2.1.1
exceptiongroup==1.2.0
fastapi==0.110.1
fastapi-mail==1.4.1
greenlet==3.0.3
h11==0.14.0
httptools==0.6.1
idna==3.7
Jinja2==3.1.3
lxml==5.2.1
MarkupSafe==2.1.5
pydantic==2.7.0
pydantic-core==2.18.1
pydantic-settings==2.2.1
pydentic==0.0.1.dev3
pypdf==4.2.0
python-docx==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
python-stdnum==1.20
PyYAML==6.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
typing-extensions==4.11.0
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
